Changing CreateAtlas obsolete signature and using Path.Combine
