+++ MenuVariantsMod/LogoLoad.cs
@@ -7,24 +7,21 @@ namespace MenuVariantsMod
     public class LogoLoad
     {
         public static List<CustomLogo> LogoList;
+        
+        private static readonly string CUSTOM_LOGOS_DIR =  Path.Combine("Mods", "MenuVariantsMod", "Content", "CustomLogos" );
+        
         public static void Load(FortContent content) {
             MenuVariantModModule.Vanilla = new List<bool>();
             MenuVariantModModule.Vanilla.Add(true);
             MenuVariantModModule.Vanilla.Add(true);
             
-            var _separator = Path.DirectorySeparatorChar.ToString();
-            string _customLogos = "Mods" + _separator + "MenuVariantsMod" + _separator + "Content" + _separator + "CustomLogos" + _separator; 
-            
-            string[] directories = Directory.GetDirectories(_customLogos);
-            string[] array = directories;
+            string[] directories = Directory.GetDirectories(CUSTOM_LOGOS_DIR);
             LogoList = new List<CustomLogo>();
-            foreach (string customLogoPath in array)
+            foreach (string customLogoPath in directories)
             {
-                Console.WriteLine(customLogoPath);
-                Atlas atlas = AtlasExt.CreateAtlas(content, customLogoPath + _separator + "atlas.xml", customLogoPath + _separator + "atlas.png", true, ContentAccess.Root);
-                var LogoData = Calc.LoadXML(customLogoPath + _separator + "LogoData.xml");
+                Atlas atlas = AtlasExt.CreateAtlas(content, Path.Combine(customLogoPath,"atlas.xml" ), Path.Combine(customLogoPath,"atlas.png"));
+                var LogoData = Calc.LoadXML(Path.Combine(customLogoPath, "LogoData.xml") );
                 var TrueLogoData = LogoData["LogoData"];
-                Console.WriteLine(atlas);
                 
                 if (atlas == null)
                 {
