+++ MenuVariantsMod/MenuVariantsModSettings.cs
